requirement.txt
torch
transformers
selenium
protobuf
sentencepiece
undetected-chromedriver
requests
jieba
numpy
scikit-learn
bertopic
sentence-transformers
hdbscan
umap-learn
pandas
matplotlib
tqdm